avences en la logica de la pagina de proveedor

+++ rfront/src/Components/PaginaProveedor/ProvAdminBar.jsx
@@ -9,15 +9,65 @@ import LogoutIcon from '@mui/icons-material/Logout';
 import DeleteForeverIcon from '@mui/icons-material/DeleteForever'; 
 import FormatColorFillIcon from '@mui/icons-material/FormatColorFill';
 import BorderColorIcon from '@mui/icons-material/BorderColor';
+import PanoramaIcon from '@mui/icons-material/Panorama';
 import Tooltip from '@mui/material/Tooltip';
+import { useNavigate } from 'react-router-dom';
+import Swal from 'sweetalert2';
 
 function ProvAdminBar({email}) {
-  const salir=()=>{
+  const navigate = useNavigate();
 
-    sessionStorage.clear()
-    window.location.href='/'
+  const handleLogOut = () => {
+    Swal.fire({
+      title: 'Está Seguro de Cerrar Sessión?',
+      text: "Para administrar su página deberá iniciar sessión nuevamente!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#004F67',
+      cancelButtonColor: '#666',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Si, cerrar sesión!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          icon: 'success',
+          title: 'Se cerró su sesión exitosamente',
+          showConfirmButton: false,
+          timer: 2000
+        })
+        sessionStorage.clear()
+        navigate('/', { replace:true })
+      }
+    })
 
   }
+
+  const handleDeletePage = () => {
+    Swal.fire({
+      title: 'Borrar Página',
+      text: "¡ADVERTENCIA! -> Esta acción no se puede revertir, ¿Está Seguro?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#666',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'No, Cancelar',
+      confirmButtonText: 'Si, Borrar Mi Página!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          icon: 'success',
+          title: 'Se eliminó su página exitosamente',
+          showConfirmButton: false,
+          timer: 2000
+        })
+        // Aqui va la petición delete
+        navigate('/publicar', { replace:true })
+      }
+    })
+  }
+
+
+
     return (
         <div className="bar-container">
             <Box sx={{ flexGrow: 1 }}>
@@ -29,12 +79,12 @@ function ProvAdminBar({email}) {
                     color="inherit"
                     aria-label="menu"
                   >
-                    {/* <MenuIcon /> */}
                     <BusinessIcon />
                   </IconButton>
                   <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>{email}</Typography>
 
-                  <IconButton
+                  <IconButton 
+                    onClick={handleDeletePage}
                     size="large"
                     aria-label="menu"
                     sx={{ color: "FireBrick" }}
@@ -45,7 +95,7 @@ function ProvAdminBar({email}) {
                   </IconButton>
 
                   <IconButton
-                    onClick={()=>salir()}
+                    onClick={handleLogOut}
                     to='/'
                     size="large"
                     color="inherit"
@@ -62,8 +112,18 @@ function ProvAdminBar({email}) {
                     color="inherit"
                     aria-label="menu"
                   >
-                    <Tooltip title="Editar Datos" arrow>
-                      <BorderColorIcon />
+                    <Tooltip title="Cambiar Colores" arrow>
+                      <FormatColorFillIcon />
+                    </Tooltip>
+                  </IconButton>
+                  
+                  <IconButton
+                    size="large"
+                    color="inherit"
+                    aria-label="menu"
+                  >
+                    <Tooltip title="Cambiar Imagen" arrow>
+                      <PanoramaIcon />
                     </Tooltip>
                   </IconButton>
 
@@ -72,8 +132,8 @@ function ProvAdminBar({email}) {
                     color="inherit"
                     aria-label="menu"
                   >
-                    <Tooltip title="Cambiar Colores" arrow>
-                      <FormatColorFillIcon />
+                    <Tooltip title="Editar Datos" arrow>
+                      <BorderColorIcon />
                     </Tooltip>
                   </IconButton>
